backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
python-multipart==0.0.6
reportlab==4.0.7
pydantic[email]==2.5.3
pydantic-settings==2.1.0
